requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alabaster=0.7.12=py_0
appdirs=1.4.4=py_0
argh=0.26.2=py38_0
astroid=2.4.2=py38_0
async_generator=1.10=py_0
atomicwrites=1.4.0=py_0
attrs=20.3.0=pyhd3eb1b0_0
autopep8=1.5.4=py_0
babel=2.9.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
bcrypt=3.2.0=py38he774522_0
black=19.10b0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2020.12.8=haa95532_0
certifi=2020.12.5=py38haa95532_0
cffi=1.14.4=py38hcd4344a_0
chardet=4.0.0=py38haa95532_1003
click=7.1.2=py_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.3.1=py38hcd4344a_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=py38_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20200713=py_0
docutils=0.16=py38_1
drlagents=0.0.0=pypi_0
entrypoints=0.3=py38_0
ffmpeg=4.3.1=ha925a31_0
flake8=3.8.4=py_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py38_1
gym=0.18.0=py38hc5df569_0
icu=58.2=ha925a31_3
idna=2.10=py_0
imagesize=1.2.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.19.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=5.6.4=py_0
jedi=0.17.2=py38haa95532_1
jinja2=2.11.2=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
keyring=21.7.0=py38haa95532_0
kiwisolver=1.3.0=py38hd77b12b_0
lazy-object-proxy=1.4.3=py38h2bbff1b_2
libblas=3.8.0=20_mkl
libcblas=3.8.0=20_mkl
liblapack=3.8.0=20_mkl
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h33f27b4_0
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
lz4-c=1.9.2=hf4a77e7_3
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38he774522_0
matplotlib=3.3.2=haa95532_0
matplotlib-base=3.3.2=py38hba9282a_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38he774522_1000
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
msys2-conda-epoch=20160418=1
mypy_extensions=0.4.3=py38_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
nest-asyncio=1.4.3=pyhd3eb1b0_0
ninja=1.10.2=py38h6d14046_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
numpydoc=1.1.0=pyhd3eb1b0_1
olefile=0.46=py_0
openssl=1.1.1i=h2bbff1b_0
packaging=20.8=pyhd3eb1b0_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.3=py38haa95532_1
paramiko=2.7.2=py_0
parso=0.7.0=py_0
pathspec=0.7.0=py_0
pathtools=0.1.2=py_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=py38h4fa10fc_0
pip=20.3.3=py38haa95532_0
pluggy=0.13.1=py38_0
prompt-toolkit=3.0.8=py_0
psutil=5.7.2=py38he774522_0
ptyprocess=0.6.0=pyhd3eb1b0_2
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pydocstyle=5.1.1=py_0
pyflakes=2.2.0=py_0
pyglet=1.5.13=py38haa244fe_0
pygments=2.7.3=pyhd3eb1b0_0
pylint=2.6.0=py38_0
pyls-black=0.4.6=hd3eb1b0_0
pyls-spyder=0.3.0=pyhd3eb1b0_0
pynacl=1.4.0=py38h62dcd97_1
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.17.3=py38he774522_0
pysocks=1.7.1=py38haa95532_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.36.2=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.7.1=py3.8_cuda101_cudnn7_0
pytz=2020.4=pyhd3eb1b0_0
pywin32=227=py38he774522_1
pywin32-ctypes=0.2.0=py38_1000
pyyaml=5.3.1=py38he774522_1
pyzmq=20.0.0=py38hd77b12b_1
qdarkstyle=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.1=py_0
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
regex=2020.11.13=py38h2bbff1b_0
requests=2.25.1=pyhd3eb1b0_0
rope=0.18.0=py_0
rtree=0.9.4=py38h21ff451_1
scipy=1.5.3=py38h5f893b4_0
setuptools=51.0.0=py38haa95532_2
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.3.0=pyhd3eb1b0_0
sphinx=3.4.0=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.2.0=py38haa95532_0
spyder-kernels=1.10.1=py38haa95532_0
sqlite=3.33.0=h2a8f88b_0
testpath=0.4.4=py_0
three-merge=0.1.1=pyhd3eb1b0_0
tk=8.6.10=he774522_0
toml=0.10.1=py_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu101
tornado=6.1=py38h2bbff1b_0
traitlets=5.0.5=py_0
typed-ast=1.4.1=py38he774522_0
typing_extensions=3.7.4.3=py_0
ujson=4.0.1=py38ha925a31_0
urllib3=1.26.2=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=0.10.4=py38haa95532_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
wrapt=1.11.2=py38he774522_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.30.0=py_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
